backend/requirements.txt
Django==3.2
django-cors-headers==3.13.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.0.4
Pillow==9.2.0
python-dotenv==0.21.0
psycopg2-binary==2.8.6
PyJWT==2.4.0
webcolors==1.11.1
django-colorfield==0.8.0
drf-extra-fields==3.4.1